Rename, format and refactor
-Rename ApplicationUser
-Change from Long to int
-Formatted Code and Imports

+++ backend/src/main/java/ch/zhaw/parkship/reservation/ReservationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<ReservationEntity, Long>{
+public interface ReservationRepository extends JpaRepository<ReservationEntity, Integer> {
 
 }
